MMS Breath (3/3) & Missed Call Breath (3/3)
This will give your SMS notifications a breathing effect in the statusbar instead of being a static icon.

-----------------

Credits: @Kroz for the icons and idea!

This will give your Missed Call notifications a breathing effect in the statusbar instead of an static icon.

Thanks to DerSchubi for the fix on CM.

Conflicts:
	res/values/strings.xml

Conflicts:
	res/values/strings.xml
	src/com/android/settings/carbon/StatusBar.java

Change-Id: I583d7c6ed9c42697e0cdb1a4fda687f6f2a81a71

Conflicts:
	res/values/cm_strings.xml
	res/xml/system_settings.xml
	src/com/android/settings/cyanogenmod/SystemSettings.java